refactor(sandboxes): move the sandbox modules out of agent/utils
The provider registry, the per-thread handle and proxy, work-dir resolution,
the transient-error retry wrapper, the reviewer's repo prep, and the read-only
backend adapter move into agent/sandboxes/, next to the providers they drive
in agent/integrations/:

  utils/sandbox.py           -> sandboxes/providers.py
  utils/sandbox_state.py     -> sandboxes/state.py
  utils/sandbox_paths.py     -> sandboxes/paths.py
  utils/sandbox_retry.py     -> sandboxes/retry.py
  utils/repo_prep.py         -> sandboxes/repo_prep.py
  utils/read_only_backend.py -> sandboxes/read_only_backend.py

agent/sandboxes/__init__.py stays empty: integrations/langsmith.py imports
providers and retry, so a package that eagerly imported its own submodules
would close a cycle the moment anything reached one of them.
