+++ agent/middleware/workflow_push_guard.py
@@ -16,6 +16,8 @@ from langgraph.prebuilt.tool_node import ToolCallRequest
 from langgraph.types import Command
 from langgraph_sdk import get_client
 
+from agent.sandboxes.state import SANDBOX_BACKENDS
+
 from ..dashboard.workflow_approval import (
     ensure_workflow_push_pending,
     mark_workflow_push_notified,
@@ -23,7 +25,6 @@ from ..dashboard.workflow_approval import (
 )
 from ..tools.slack_thread_reply import build_workflow_approval_blocks
 from ..utils.dashboard_links import dashboard_workflow_approval_url
-from ..utils.sandbox_state import SANDBOX_BACKENDS
 from ..utils.slack import (
     LANGGRAPH_URL,
     get_active_slack_thread,
