+++ agent/reviewer.py
@@ -39,6 +39,10 @@ from langchain.agents.middleware import ModelCallLimitMiddleware
 from langchain.agents.middleware.types import AgentMiddleware
 from langchain_core.language_models.chat_models import BaseChatModel
 
+from agent.sandboxes.paths import aresolve_sandbox_work_dir
+from agent.sandboxes.repo_prep import materialize_trusted_skills, prepare_review_repo
+from agent.sandboxes.state import SandboxUnreachableError
+
 from .dashboard.options import gate_fable_model
 from .dashboard.team_settings import (
     get_effective_gateway_enabled,
@@ -114,9 +118,6 @@ from .utils.deferred_model import make_deferred_error_model
 from .utils.github_app import get_github_app_installation_token_with_expiry
 from .utils.github_token import cache_github_token_for_thread
 from .utils.model import DEFAULT_LLM_REASONING, make_model, provider_model_kwargs
-from .utils.repo_prep import materialize_trusted_skills, prepare_review_repo
-from .utils.sandbox_paths import aresolve_sandbox_work_dir
-from .utils.sandbox_state import SandboxUnreachableError
 from .utils.tracing import REVIEW_TRACING_PROJECT, traced_graph_factory
 
 HISTORICAL_REVIEW_GUIDANCE = """- **Anything that overlaps an existing PR review thread.** A
