+++ agent/tools/background_execute.py
@@ -10,7 +10,7 @@ from typing import Any, Literal
 
 from langgraph.config import get_config
 
-from ..utils.sandbox_state import SANDBOX_BACKENDS
+from agent.sandboxes.state import SANDBOX_BACKENDS
 
 logger = logging.getLogger(__name__)
