+++ agent/tools/create_sandbox_service_url.py
@@ -3,7 +3,7 @@ from typing import Any
 from langgraph.config import get_config
 
 from agent.integrations.langsmith import get_async_sandbox_client
-from agent.utils.sandbox_state import get_sandbox_backend, unwrap_sandbox_backend
+from agent.sandboxes.state import get_sandbox_backend, unwrap_sandbox_backend
 
 
 async def create_sandbox_service_url(
