+++ agent/tools/save_plan.py
@@ -8,6 +8,8 @@ from typing import Annotated, Any
 from langgraph.config import get_config
 from langgraph.prebuilt import InjectedState
 
+from agent.sandboxes.state import get_sandbox_backend
+
 from ..dashboard.plan_store import (
     PLAN_FILE_DIRECTORY,
     PLAN_STATUS_READY,
@@ -15,7 +17,6 @@ from ..dashboard.plan_store import (
     save_plan_content,
 )
 from ..utils.html_artifact import DEFAULT_TITLE, wrap_html_artifact
-from ..utils.sandbox_state import get_sandbox_backend
 
 logger = logging.getLogger(__name__)
