+++ tests/agent/test_factory_tool_loading.py
@@ -12,8 +12,8 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 from langgraph.graph.state import RunnableConfig
 
+from agent.sandboxes.state import SANDBOX_BACKENDS
 from agent.server import get_agent
-from agent.utils.sandbox_state import SANDBOX_BACKENDS
 
 _START_TIMEOUT_SECONDS = 2.0
