+++ tests/middleware/test_subdir_agents_middleware.py
@@ -7,7 +7,7 @@ import pytest
 from langchain_core.messages import ToolMessage
 
 from agent.middleware.subdir_agents import SubdirAgentsReadMiddleware
-from agent.utils import sandbox_state
+from agent.sandboxes import state as sandbox_state
 
 
 class FakeReadResult:
