+++ tests/sandbox/test_reviewer_sandbox_recovery.py
@@ -12,9 +12,9 @@ from langchain_core.runnables import RunnableConfig
 from langsmith.sandbox import SandboxClientError
 
 from agent.reviewer import PrepareReviewerRunMiddleware, _ensure_reviewer_sandbox_for_thread
+from agent.sandboxes.providers import SandboxGoneError
+from agent.sandboxes.state import SandboxUnreachableError, set_sandbox_backend
 from agent.server import SANDBOX_BACKENDS, ensure_sandbox_for_thread
-from agent.utils.sandbox import SandboxGoneError
-from agent.utils.sandbox_state import SandboxUnreachableError, set_sandbox_backend
 
 
 @pytest.mark.asyncio
