+++ tests/sandbox/test_sandbox_recreation.py
@@ -2,8 +2,8 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
+from agent.sandboxes.state import SANDBOX_BACKENDS, set_sandbox_backend
 from agent.server import recreate_sandbox_for_thread
-from agent.utils.sandbox_state import SANDBOX_BACKENDS, set_sandbox_backend
 
 
 @pytest.mark.asyncio
