+++ tests/sandbox/test_sandbox_retry.py
@@ -3,7 +3,7 @@ from unittest.mock import AsyncMock, patch
 import pytest
 from langsmith.sandbox import SandboxClientError, SandboxRetryableConnectionError
 
-from agent.utils.sandbox_retry import (
+from agent.sandboxes.retry import (
     MAX_TRANSIENT_ATTEMPTS,
     is_transient_sandbox_error,
     retry_transient_sandbox_errors,
@@ -32,7 +32,7 @@ async def test_a_gateway_blip_is_retried_until_it_clears() -> None:
             raise _transient()
         return "ok"
 
-    with patch("agent.utils.sandbox_retry.asyncio.sleep", new_callable=AsyncMock) as mock_sleep:
+    with patch("agent.sandboxes.retry.asyncio.sleep", new_callable=AsyncMock) as mock_sleep:
         result = await retry_transient_sandbox_errors(operation, description="test")
 
     assert result == "ok"
@@ -58,7 +58,7 @@ async def test_retries_are_bounded() -> None:
     operation = AsyncMock(side_effect=_transient())
 
     with (
-        patch("agent.utils.sandbox_retry.asyncio.sleep", new_callable=AsyncMock),
+        patch("agent.sandboxes.retry.asyncio.sleep", new_callable=AsyncMock),
         pytest.raises(SandboxRetryableConnectionError),
     ):
         await retry_transient_sandbox_errors(operation, description="test")
